+++ src/main/java/com/example/capstoneproject/service/impl/InvolvementServiceImpl.java
@@ -244,7 +244,11 @@ public class InvolvementServiceImpl extends AbstractBaseService<Involvement, Inv
                 scoreRepository.deleteScoreById(score.getId());
             }
 
-            sectionLogRepository.deleteBySection_Id(section.getId());
+            if(section!=null){
+                sectionLogRepository.deleteBySection_Id(section.getId());
+                cv.setOverview(null);
+                cvRepository.save(cv);
+            }
             //Get process evaluate
             List<BulletPointDto> evaluateResult = evaluateService.checkSentences(dto.getDescription());
             InvolvementViewDto involvementViewDto = new InvolvementViewDto();
@@ -265,7 +269,7 @@ public class InvolvementServiceImpl extends AbstractBaseService<Involvement, Inv
                 Evaluate evaluate = evaluates.get(i);
                 BulletPointDto bulletPointDto = evaluateResult.get(i);
                 SectionLogDto sectionLogDto1 = new SectionLogDto();
-                sectionLogDto1.setSection(sectionMapper.mapDtoToEntity(sectionMapper.mapEntityToDto(section)));
+                sectionLogDto1.setSection(section);
                 sectionLogDto1.setEvaluate(evaluate);
                 sectionLogDto1.setBullet(bulletPointDto.getResult());
                 sectionLogDto1.setStatus(bulletPointDto.getStatus());
